add version package
